+++ docs/guide/changelog.md
@@ -4,6 +4,11 @@
 
 对于新功能、新组件、`bug` 修复以及文档更新，您可以向 `main` 分支创建拉取请求或通过右下角邮箱地址联系我
 
+## <VersionDateTag date="2025-03-17">2.3.3</VersionDateTag>
+
+- 组件库中包含 `VNode` 类型节点渲染的组件，添加 `isVNode` 判断
+- 组件库及文档代码优化
+
 ## <VersionDateTag date="2025-03-15">2.3.2</VersionDateTag>
 
 - 优化并更新 [按钮 Button](https://themusecatcher.github.io/vue-amazing-ui/guide/components/button.html) 组件，删除 `rippleColor` 属性，新增 `color` 属性，支持自定义按钮颜色